Moved functions in_notebook() and select_pbar_cls() as well as class SimpleProgressBar to file helper.py
