+++ raisr/raisr.py
@@ -3,75 +3,9 @@ import numpy as np
 import scipy.sparse.csgraph
 import pickle
 import matplotlib.pyplot as plt
-import sys
-import os
 from math import floor
 from .image import Image
-from .helper import make_slice_list
-
-try:
-    import tqdm
-    TQDM_AVAILABLE = True
-except ImportError:
-    TQDM_AVAILABLE = False
-
-
-def in_notebook():
-    if 'ipykernel' in sys.modules:
-        if any('SPYDER' in name for name in os.environ):
-            return False
-        if any('PYCHARM' in name for name in os.environ):
-            return False
-        return True
-    else:
-        return False
-
-
-def select_pbar_cls():
-    if TQDM_AVAILABLE:
-        if in_notebook():
-            return tqdm.tqdm_notebook
-        else:
-            return tqdm.tqdm
-    else:
-        return SimpleProgressBar
-
-
-class SimpleProgressBar:
-    def __init__(self, total = 100, desc = "", **kwargs):
-        self._total = total
-        self._desc = desc
-        self._count = 0
-    
-    def __enter__(self):
-        self.open()
-        return self
-    
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.close()
-        return False
-    
-    def __del__(self):
-        self.close()
-    
-    def open(self):
-        if len(self._desc) > 0:
-            print(self._desc)
-    
-    def close(self):
-        print('')
-    
-    def update(self, n = 1):
-        old_count = self._count
-        self._count += n
-        new_count = old_count + n
-        total = self._total
-        if (new_count * 100) // total != (old_count * 100) // total:
-            print('\r|', end='')
-            print('#' * ((new_count * 100) // (2 * total)), end='')
-            print(' ' * (50 - (new_count * 100) // (2 * total)), end='')
-            print('|  {0}%'.format((new_count * 100) // total), end='')
-
+from .helper import select_pbar_cls, make_slice_list
 
 class RAISR:
     def __init__(self, *, ratio = 2, patchsize = 11, gradientsize = 9,
@@ -427,10 +361,6 @@ class RAISR:
 
     def _make_pbar_kwargs(self, total = 100, desc = ""):
         kwargs = {'total': total, 'desc': desc}
-#        if total > 1e6:
-#            kwargs['mininterval'] = 1
-#        else:
-#            kwargs['mininterval'] = 0.1
         if self._pbar_cls.__name__ == 'tqdm':
             kwargs['bar_format'] =  '{desc}: {percentage:3.0f}%|{bar}| [{elapsed} elapsed/{remaining} remaining]'
         if self._pbar_cls.__name__ == 'tqdm_notebook':
